refactor: move init functions to new module
main.rs was a bit cluttered, so this helped tidy up some things.

I removed some unused fields in the bpf headers since ancestor level was
helpful for debugging, but not anymore.

Instead of having a large callback variable in main, this has been
extracted to a function that returns the proper callback.